changed SemplestBiddingBudgetMultFactor to Double
git-svn-id: file:///semplestb/final_repo@3220 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SemplestServices/src/semplest/service/bidding/BidGeneratorObj.java
@@ -74,7 +74,7 @@ public class BidGeneratorObj
 	private Long stepRest;// = 800000L;
 	
 	Double googlePercent;
-	Float budgetFactor;
+	Double budgetFactor;
 	
 
 	private String googleAccountID = null;
@@ -154,7 +154,7 @@ public class BidGeneratorObj
 		
 		final Integer googlePercentInteger = (Integer)SemplestConfiguration.configData.get("SemplestBiddingGooglePercent");		
 		googlePercent =  Double.valueOf("" + googlePercentInteger);
-		budgetFactor =  (Float) SemplestConfiguration.configData.get("SemplestBiddingBudgetMultFactor");
+		budgetFactor =  (Double) SemplestConfiguration.configData.get("SemplestBiddingBudgetMultFactor");
